Add a tool to generate fluent files
I am looking to add a benchmark for parsing large fluent files, in order to
test potential performance optimisations. Originally I used an actual
fluent file from my project, but that cannot be shared.

This change adds a script which can be used to generate "random" fluent
files, which share similar statistical properties to the actual file I
used. For example, the ratio of comments to messages, or the number of
elements in side a message.