Add a benchmark for compiling a file with 10000 items
Prior to this change, all of the benchmarks were micro benchmarks.
However, I work with a project that contains some large fluent files,
and we have observed some slowness when compiling them.

This change adds a benchmark for compiling a 10000 item file, which can
be used to verify performance optimisations in other commits.