Add queries to check is a name is reserved
Prior to this change, code which was checking whether a name was
reserved in a given scope would need to fetch the relevant set of names
and check whether the name was in it directly.

This change adds query functions which allow the checks to be performed
directly. The query functions recurse to the parent scope, rather than
needing to fetch a new set containing this scope's names and the parent
scope's.