Use keyword.iskeyword rather than expanding the set of reserved names
Prior to this change, if we were checking whether a non-builtin name was
allowed, we would expand the set of reserved names to include all
keywords.

This change switches to using the iskeyword function instead, which will
allow us to stop using the set of reserved names altogether in a future
commit.