Remove set generation methods from Scope
Prior to this change, we had replaced all callers with the query
equivalents. This provides a considerable performance improvement when
compiling large files, as we are no longer repeatedly creating new sets.

This change removes the unused methods which generated the sets as they
are no longer needed (and are a footgun).